feat: API Model을 View에 반영
